change: Rename collectioncounts module to CollectionCounts
