+++ src/web/pages/alerts/__tests__/DetailsPage.test.jsx
@@ -5,7 +5,7 @@
 
 import {describe, test, expect, testing} from '@gsa/testing';
 import Capabilities from 'gmp/capabilities/capabilities';
-import CollectionCounts from 'gmp/collection/collectioncounts';
+import CollectionCounts from 'gmp/collection/CollectionCounts';
 import Alert from 'gmp/models/alert';
 import Filter from 'gmp/models/filter';
 import {currentSettingsDefaultResponse} from 'web/pages/__mocks__/CurrentSettings';
