+++ src/web/pages/credentials/__tests__/ListPage.test.jsx
@@ -5,7 +5,7 @@
 
 import {describe, test, expect, testing} from '@gsa/testing';
 import Capabilities from 'gmp/capabilities/capabilities';
-import CollectionCounts from 'gmp/collection/collectioncounts';
+import CollectionCounts from 'gmp/collection/CollectionCounts';
 import Credential from 'gmp/models/credential';
 import Filter from 'gmp/models/filter';
 import {
