+++ src/web/pages/hosts/__tests__/DetailsPage.test.jsx
@@ -4,7 +4,7 @@
  */
 
 import {describe, test, expect, testing} from '@gsa/testing';
-import CollectionCounts from 'gmp/collection/collectioncounts';
+import CollectionCounts from 'gmp/collection/CollectionCounts';
 import Filter from 'gmp/models/filter';
 import Host from 'gmp/models/host';
 import {SEVERITY_RATING_CVSS_3} from 'gmp/utils/severity';
