+++ src/web/pages/reportconfigs/__tests__/ListPage.test.jsx
@@ -5,7 +5,7 @@
 
 import {describe, test, expect, testing} from '@gsa/testing';
 import Capabilities from 'gmp/capabilities/capabilities';
-import CollectionCounts from 'gmp/collection/collectioncounts';
+import CollectionCounts from 'gmp/collection/CollectionCounts';
 import Filter from 'gmp/models/filter';
 import ReportConfig from 'gmp/models/reportconfig';
 import {clickElement, testBulkTrashcanDialog} from 'web/components/testing';
