+++ src/web/pages/results/__tests__/ListPage.test.jsx
@@ -4,7 +4,7 @@
  */
 
 import {describe, test, expect, testing} from '@gsa/testing';
-import CollectionCounts from 'gmp/collection/collectioncounts';
+import CollectionCounts from 'gmp/collection/CollectionCounts';
 import Filter from 'gmp/models/filter';
 import Result from 'gmp/models/result';
 import {SEVERITY_RATING_CVSS_3} from 'gmp/utils/severity';
